fix(back): fix migration file for sqlite
